refactor: MenuController, MenuService 기능 분리
