+++ src/main/java/menu/Application.java
@@ -1,13 +1,19 @@
 package menu;
 
-import menu.controller.Controller;
+import menu.controller.MenuController;
+import menu.service.MenuService;
 import menu.view.InputView;
 import menu.view.OutputView;
 
 
 public class Application {
     public static void main(String[] args) {
-        Controller controller = new Controller(new InputView(), new OutputView());
-        controller.execute();
+        MenuController menuController = new MenuController(
+                new InputView(),
+                new OutputView(),
+                new MenuService()
+        );
+
+        menuController.execute();
     }
 }
